test(profiles): fold the severity guard into the mirror test + strictest-wins defaults
Panel r4 findings.

The severity suite lived in its own profile-severity.test.ts while
tests/profiles.test.ts already covered src/config/profiles.ts, leaving two test
files for one source and an orphan under the test-file-mirrors-source rule this
repo ships. Merged into profiles.test.ts and the separate file removed. The
existing "no rule is forced off in opinionated" loop went with it — the merged
suite asserts every STRUCTURE_RULE is exactly "error" there, which subsumes it.

packDefaults() indexed by bare rule name with last-write-wins, so if two packs
ever shipped the same rule name a weaker later default could mask an override
that lowers the rule below an earlier pack's stricter one. It now keeps the
strictest default per name. No duplicates exist today (117 rules, 0 collisions)
and buildPackEslintConfig throws on collision among ACTIVE packs, but this helper
walks every pack, so it should not depend on that.

Both guards re-verified by mutation: adding a rule to STRUCTURE_RULES, and
setting max-hooks-per-file back to "warn", each fail three tests.
